Modal message when there is firebase error during saving a checklist and when user tries edit a checklist.

+++ src/app/shared/checklist.service.ts
@@ -52,6 +52,8 @@ export class ChecklistService {
   }
 
   selectChecklistItem(item: ChecklistItem) {
+    if (!this.selectedIdsReference) { return; }
+
     if (item.subitems.length === 0) {
       this.selectedIdsReference
         .set({ [item.id]: true }, { merge: true });
@@ -63,6 +65,8 @@ export class ChecklistService {
   }
 
   unselectChecklistItem(item: ChecklistItem) {
+    if (!this.selectedIdsReference) { return; }
+
     if (item.subitems.length === 0) {
       this.selectedIdsReference
         .set({ [item.id]: firebase.firestore.FieldValue.delete() }, { merge: true });
